(PC-19656)[PRO] refactor: move summary components and delete empty folder Offer
