+++ pro/src/screens/OfferIndividual/Summary/DisplayOfferInAppLink/DisplayOfferInAppLink.tsx
