+++ pro/src/screens/OfferIndividual/Summary/DisplayOfferInAppLink/index.ts
