+++ pro/src/screens/OfferIndividual/Summary/Summary.tsx
@@ -22,15 +22,15 @@ import { IOfferSubCategory } from 'core/Offers/types'
 import { getOfferIndividualUrl } from 'core/Offers/utils/getOfferIndividualUrl'
 import { useNavigate, useOfferWizardMode } from 'hooks'
 import useAnalytics from 'hooks/useAnalytics'
+import useNewOfferCreationJourney from 'hooks/useNewOfferCreationJourney'
 import useNotification from 'hooks/useNotification'
 import { IcoParty } from 'icons'
 import { ReactComponent as PhoneInfo } from 'icons/info-phone.svg'
-import { DisplayOfferInAppLink } from 'pages/Offers/Offer/DisplayOfferInAppLink'
-import OfferStatusBanner from 'pages/Offers/Offer/OfferDetails/OfferStatusBanner'
+import { DisplayOfferInAppLink } from 'screens/OfferIndividual/Summary/DisplayOfferInAppLink'
+import OfferStatusBanner from 'screens/OfferIndividual/Summary/OfferStatusBanner'
 import { ButtonVariant } from 'ui-kit/Button/types'
 import { getOfferConditionalFields } from 'utils/getOfferConditionalFields'
 
-import useNewOfferCreationJourney from '../../../hooks/useNewOfferCreationJourney'
 import { ActionBar } from '../ActionBar'
 import { SynchronizedProviderInformation } from '../SynchronisedProviderInfos'
